Fix the dependency issues in CMake file.
dflowmap does not depend on expression optimization tools to work (although it is recommended to have the tool installed for optimization purpose), so we fix the dependency issues in the cmake file.

Also, we assume that users do NOT have the expression optimization tools installed by default, so we remove the pre-configurations in config.in.